+++ app/Http/Controllers/API/TypeCampaignController.php
@@ -3,7 +3,13 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\TypeCampaign\CreateTypeCampaignRequest;
+use App\Http\Requests\TypeCampaign\UpdateTypeCampaignRequest;
+use App\Http\Resources\Campaign\CampaignResource;
+use App\Http\Resources\TypeCampaign\TypeCampaignResource;
+use App\Models\TypeCampaign;
 use App\Services\Interfaces\TypeCampaignServiceInterface;
+use http\Client\Curl\User;
 use Illuminate\Http\Request;
 use Laravel\Telescope\AuthorizesRequests;
 
@@ -20,33 +26,48 @@ class TypeCampaignController extends Controller
 
     public function index()
     {
-        return $this->typeCampaignService->getAllTypeCampaign();
+        $this->authorize('viewAny', TypeCampaign::class);
+        $typeCampaign = $this->typeCampaignService->getAllTypeCampaign();
+
+        return  TypeCampaignResource::collection($typeCampaign);
     }
 
     public function show(int $id)
     {
-        return $this->typeCampaignService->getTypeCampaignById($id);
+        $this->authorize('view', $this->typeCampaignService->getTypeCampaignById($id));
+        $typeCampaign = $this->typeCampaignService->getTypeCampaignById($id);
+        return new TypeCampaignResource($typeCampaign);
     }
 
-    public function store(Request $request)
+    public function store(CreateTypeCampaignRequest $request)
     {
-        $data = $request->all();
-        return $this->typeCampaignService->createTypeCampaign($data);
+        $data = $request->validated();
+        $this->authorize('create', TypeCampaign::class);
+        $typeCampaign = $this->typeCampaignService->createTypeCampaign($data);
+        return new TypeCampaignResource($typeCampaign);
     }
 
-    public function update(Request $request, int $id)
+    public function update( UpdateTypeCampaignRequest $request, int $id)
     {
-        $data = $request->all();
-        return $this->typeCampaignService->updateTypeCampaign($id, $data);
+        $data = $request->validated();
+        $typeCampaign = $this->typeCampaignService->getTypeCampaignById($id);
+        $this->authorize('update', $typeCampaign);
+        $typeCampaign = $this->typeCampaignService->updateTypeCampaign($id, $data);
+        return  new TypeCampaignResource($typeCampaign);
     }
 
     public function destroy(int $id)
     {
-        return $this->typeCampaignService->deleteTypeCampaign($id);
+        $typeCampaign = $this->typeCampaignService->getTypeCampaignById($id);
+        $this->authorize('delete', $typeCampaign);
+        $this->typeCampaignService->deleteTypeCampaign($id);
+
+        return response()->json(['message' => 'Supprimé avec succès.'], 200);
     }
 
     public function showByCriteria(Request $request)
     {
-        return $this->typeCampaignService->getTypeCampaignByCriteria($request->query());
+        $typeCampaign = $this->typeCampaignService->getTypeCampaignByCriteria($request->query());
+        return new TypeCampaignCollection($typeCampaign);
     }
 }
